+++ components/screens/programming/tabs/MixedViewTab.js
