+++ stores/SpeedsStore.js
